Move attachments accordion to its own component
