+++ src/components/DetailView/Components/SplitIssue/SplitIssueCreate.tsx
@@ -4,7 +4,7 @@ import { showNotification } from "@mantine/notifications";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { Issue } from "types";
 import { useCanvasStore } from "@canvas/lib/Store";
-import { getResource, uploadAttachment } from "../Attachments/queryFunctions";
+import { uploadAttachment } from "../../../issue-details/Attachments/queryFunctions";
 import {
   IssueTypeSelect,
   StatusSelect,
@@ -77,7 +77,7 @@ export function SplitIssueCreate({
       const filesForm = new FormData();
       if (files) {
         files.forEach((f) => filesForm.append("file", f, f.name));
-        getResource().then((r) => uploadAttachment(issueKey, r, filesForm));
+        window.provider.getResource().then((r) => uploadAttachment(issueKey, r, filesForm));
       }
       showNotification({
         message: `The issue ${issueKey} has been created!`,
