+++ src/components/issue-details/Attachments/index.ts
@@ -1,0 +1,2 @@
+export * from "./AttachmentsAccordion";
+export * from "./Attachments";
